# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_alemancenter_fiber-api_internal_models.SecurityLog:
    properties:
      attack_type:
        type: string
      city:
        type: string
      country_code:
        type: string
      created_at:
        type: string
      description:
        type: string
      event_type:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      is_blocked:
        type: boolean
      is_resolved:
        type: boolean
      is_trusted:
        type: boolean
      method:
        type: string
      request_data:
        type: string
      resolved_at:
        type: string
      risk_score:
        type: integer
      route:
        type: string
      severity:
        type: string
      updated_at:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_alemancenter_fiber-api_internal_repositories.TopAttackingIP:
    properties:
      count:
        type: integer
      ip_address:
        type: string
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.IPDetailsResponse:
    properties:
      ip:
        type: string
      is_blocked:
        type: boolean
      is_trusted:
        type: boolean
      recent_logs:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.SecurityLog'
        type: array
      total_events:
        type: integer
    type: object
...
  github_com_alemancenter_fiber-api_internal_services.SecurityOverviewResponse:
    properties:
      last_7d_events:
        type: integer
      last_24h_events:
        type: integer
      top_ips:
        items:
          $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_repositories.TopAttackingIP'
        type: array
      total_attacks:
        type: integer
    type: object
  github_com_alemancenter_fiber-api_internal_services.SecurityStatsResponse:
    properties:
      blocked_ips:
        type: integer
      critical_logs:
        type: integer
      resolved_logs:
        type: integer
      total_logs:
        type: integer
      trusted_ips:
        type: integer
    type: object
...
  github_com_alemancenter_fiber-api_internal_utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      pagination:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      success:
        type: boolean
      token:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_utils.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handlers_security.ipPayload:
    properties:
      ip:
        type: string
      ip_address:
        type: string
      note:
        type: string
      reason:
        type: string
    type: object
...
host: api.alemancenter.com
info:
  contact:
    email: support@alemancenter.com
    name: API Support
  description: Backend API for Alemancenter Educational Platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Alemancenter API
  version: 2.0.0
paths:
...
  /dashboard/security/ip/{ip}:
    get:
      description: Returns details, status, and recent logs for a specific IP address
      parameters:
      - description: IP Address
        in: path
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.IPDetailsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get IP Details
      tags:
      - Security
  /dashboard/security/ip/{ip}/block:
    post:
      consumes:
      - application/json
      description: Block a specific IP address
      parameters:
      - description: IP Address
        in: path
        name: ip
        required: true
        type: string
      - description: Optional payload (reason, note)
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_handlers_security.ipPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Block IP
      tags:
      - Security
  /dashboard/security/ip/{ip}/trust:
    post:
      consumes:
      - application/json
      description: Mark a specific IP address as trusted (bypassing certain limits)
      parameters:
      - description: IP Address
        in: path
        name: ip
        required: true
        type: string
      - description: Optional payload (note)
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_handlers_security.ipPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Trust IP
      tags:
      - Security
  /dashboard/security/ip/{ip}/unblock:
    post:
      description: Unblock a previously blocked IP address
      parameters:
      - description: IP Address
        in: path
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Unblock IP
      tags:
      - Security
  /dashboard/security/ip/{ip}/untrust:
    post:
      description: Remove a specific IP address from the trusted list
      parameters:
      - description: IP Address
        in: path
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Untrust IP
      tags:
      - Security
  /dashboard/security/logs:
    delete:
      description: Delete all security logs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Clear All Logs
      tags:
      - Security
    get:
      description: Returns paginated security logs with optional filters
      parameters:
      - description: Filter by severity (low, medium, high, critical)
        in: query
        name: severity
        type: string
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by IP address
        in: query
        name: ip
        type: string
      - description: Filter by resolution status (true/false)
        in: query
        name: resolved
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_models.SecurityLog'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List Security Logs
      tags:
      - Security
  /dashboard/security/logs/{id}:
    delete:
      description: Delete a specific security log entry by ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Security Log
      tags:
      - Security
  /dashboard/security/logs/{id}/resolve:
    post:
      description: Marks a specific security log entry as resolved
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Resolve Security Log
      tags:
      - Security
  /dashboard/security/monitor/dashboard:
    get:
      description: Returns a comprehensive payload (stats, recent logs) for the frontend
        monitor view
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Security Monitor Dashboard
      tags:
      - Security
  /dashboard/security/overview:
    get:
      description: Returns an overview of security metrics (last 24h, 7d, total attacks,
        and top IPs)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.SecurityOverviewResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Security Overview
      tags:
      - Security
  /dashboard/security/stats:
    get:
      description: Get overall security statistics including total logs, critical
        logs, and blocked IPs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.SecurityStatsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Security Statistics
      tags:
      - Security
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  FrontendKeyAuth:
    description: Frontend identifier key for public endpoints.
    in: header
    name: X-Frontend-Key
    type: apiKey
swagger: "2.0"